Refactor logic file into different files
